+++ src/models/Module.java
@@ -1,10 +1,10 @@
 public class Module {
 
 	private String name;
-	private int estimate;
+	private double estimate;
 	private String developmentMethod;
 
-	public Module(String name, int estimate, String developmentMethod) {
+	public Module(String name, double estimate, String developmentMethod) {
 		this.name = name;
 		this.estimate = estimate;
 		this.developmentMethod = developmentMethod;
@@ -14,7 +14,7 @@ public class Module {
 		return name;
 	}
 
-	public int getEstimate() {
+	public double getEstimate() {
 		return estimate;
 	}
